# openapi/mfethuls-mvp.yaml
openapi: 3.1.0
info:
  title: MFethuls Ingest & Query API (MVP)
  version: "0.1.0"
  description: Minimal OpenAPI spec for registry preview, manual ingest, jobs, datasets and queries.
servers:
  - url: https://api.example.com/v1
paths:
  /registry/preview:
    post:
      summary: Parse and preview a registry spreadsheet without ingesting
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PreviewUpload'
      responses:
        '200':
          description: Preview with validation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreviewResponse'
  /ingest:
    post:
      summary: Start ingestion job (manual trigger)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/IngestUpload'
      responses:
        '202':
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
  /jobs/{job_id}:
    get:
      summary: Get job status
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
  /datasets:
    get:
      summary: List datasets
      parameters:
        - in: query
          name: project_id
          schema: { type: string }
      responses:
        '200':
          description: Dataset list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DatasetSummary'
  /queries:
    post:
      summary: Execute SQL against registered datasets/views
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: Synchronous result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '202':
          description: Async job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
  /health:
    get:
      summary: Liveness
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
components:
  schemas:
    PreviewUpload:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
    PreviewResponse:
      type: object
      properties:
        rows:
          type: array
          items:
            type: object
            properties:
              row_number:
                type: integer
              values:
                type: object
              valid:
                type: boolean
              warnings:
                type: array
                items:
                  type: object
                  properties:
                    field: { type: string }
                    message: { type: string }
    IngestUpload:
      type: object
      properties:
        file:
          type: string
          format: binary
        storage_mode:
          type: string
          enum: [local, cloud, both]
        cloud_provider:
          type: string
          enum: [s3, azure]
    JobResponse:
      type: object
      properties:
        job_id:
          type: string
        status:
          type: string
          enum: [queued, running, completed, failed]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        message:
          type: string
        registry_storage_path:
          type: string
        registry_table:
          type: string
        datasets:
          type: array
          items:
            type: object
            properties:
              experiment_id: { type: string }
              dataset_id: { type: string }
              storage_path: { type: string }
              status: { type: string }
    DatasetSummary:
      type: object
      properties:
        dataset_id: { type: string }
        name: { type: string }
        storage_mode: { type: string }
        queryable: { type: boolean }
    QueryRequest:
      type: object
      required: [project_id, sql]
      properties:
        project_id: { type: string }
        dataset_ids:
          type: array
          items: { type: string }
        sql: { type: string }
        mode:
          type: string
          enum: [sync, async]
          default: sync
    QueryResponse:
      type: object
      properties:
        query_id: { type: string }
        status:
          type: string
          enum: [completed, running, failed]
        columns:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              type: { type: string }
        rows:
          type: array
          items:
            type: array
            items: {}